Implement Repo, UoW and related data services
Implemented the Repository, Unit of Work, Repository Provider and
Repository Factories for the Raffles solution. This is a simple first
pass, and will likely require additional implementations.

* Slight code reorganization - removed EF from the namespaces and code
structure.

* Repository Factories take a DbContext and return a repository bound to
the DbContext. This separates concrete repositories out of the
UnitOfWork.

* Repository Provider is more geared towards client/server interaction
and is likely not needed for this project. However, I wanted to check it
out. Helps keep concrete repositories separate from the UnitOfWork.

* UnitOfWork separates the DbContext and Repositories from the view
models, commits changes to the database.

* Repository classes - generic repository for standard entity framework
entities, and specific repositories for special querying needs.